+++ utils/model.py
@@ -1,0 +1,6 @@
+users:list=[
+    {'name': 'Dobrawa','location':'Warszawa','posts':100},
+    {'name': 'Patrycja','location':'Kutno','posts':200},
+    {'name': 'Maja','location':'Inowrocław','posts':300},
+    {'name': 'Mateusz','location':'Żeronice','posts':400},
+]
